Extract commonalities from Phone, Email, Address Create Contact class Add String value parameter to Address.
