+++ src/seedu/addressbook/storage/jaxb/AdaptedPerson.java
@@ -66,7 +66,7 @@ public class AdaptedPerson {
 
         address = new AdaptedContactDetail();
         address.isPrivate = source.getAddress().isPrivate();
-        address.value = source.getAddress().toString();
+        address.value = source.getAddress().value;
 
         tagged = new ArrayList<>();
         for (Tag tag : source.getTags()) {
