+++ test/java/seedu/addressbook/parser/ParserTest.java
@@ -288,7 +288,7 @@ public class ParserTest {
                 + person.getName().fullName
                 + (person.getPhone().isPrivate() ? " pp/" : " p/") + person.getPhone().value
                 + (person.getEmail().isPrivate() ? " pe/" : " e/") + person.getEmail().value
-                + (person.getAddress().isPrivate() ? " pa/" : " a/") + person.getAddress().toString();
+                + (person.getAddress().isPrivate() ? " pa/" : " a/") + person.getAddress().value;
         for (Tag tag : person.getTags()) {
             addCommand += " t/" + tag.tagName;
         }
